\name{TestScorer-internal}
\alias{add.new.test}
\alias{delete.test}
\alias{make.scale.window}
\alias{make.test.window}
\alias{setTestsDirectory}
\alias{write.scale}
\alias{write.test.begining}
\alias{write.test.end}
\alias{testChar}
\alias{scoring.fun}
\title{Internal mypackage Functions}
\description{
  Internal mypackage functions
}
\details{
  These are not to be called by the user.
}
\keyword{ internal }