\name{TestScorer-package}
\alias{TestScorer-package}
\docType{package}
\title{Scoring tests and questionnaires}
\description{
The main function allows the user to introduce the items of a test or questionnaire and obtain raw and transformed scores of scales. Also allows adding new questionaire scoring instructions.
}
\details{
\tabular{ll}{
Package: \tab TestScorer\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2013-11-30\cr
License: \tab GPL-2\cr
}
Main function: \code{\link{TestScorerGUI}}
}
\author{
Manel Salamero

Maintainer: Manel Salamero <manelsalamero@gmail.com>
}
\keyword{ package }
