\name{TestScorerGUI}

\alias{TestScorerGUI}

\title{Graphical user interface to score tests and questionnaires (raw and transformed scores)}

\description{This function allows the user to introduce the items of a test or questionnaire and obtain raw and transformed scores of scales. Also allows adding new questionaire scoring directives.}

\usage{
TestScorerGUI()
}

\arguments{The function don not admit arguments.
}

\details{
When first running the funcion a directory for scoring tests is created. Some tests examples and the document TestScorerInstructions.pdf are copied to the directory.
}

\value{
No value is returned. The function is used for its side effects.}

\author{
Manel Salamero
}

\note{
See the document TestScorerInstruction.pdf. It explain the way of introducing and scoring the answers of a test and the procedure for creating new ones.
}

\examples{
\dontrun{
TestScorerGUI()
}
}

\keyword{utilities}
